requirements.txt
streamlit~=1.27.2
anthropic~=0.3.11
python-dotenv~=1.0.0